refactor: modal 타입 리팩토링
