+++ client/src/lib/redux/reducer/index.ts
@@ -2,7 +2,7 @@ import { PayloadAction, combineReducers } from '@reduxjs/toolkit';
 import { authSlice } from '../slices/authSlice';
 import { modalSlice } from '../slices/modalSlice';
 
-export const reducer = (state: any, action: PayloadAction<any>) => {
+export const reducer = (state: any, action: PayloadAction) => {
   return combineReducers({
     auth: authSlice.reducer,
     modal: modalSlice.reducer,
