+++ client/src/pages/board/index.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/components/Layout';
+import { Layout } from '@/components';
 import React from 'react';
 
 export default function Board() {
