+++ client/src/types/redux/IModal.ts
@@ -1,0 +1,7 @@
+import { ModalType } from '@/constants/constant';
+
+export interface IModal {
+  type: ModalType;
+  message: string;
+  callback?: (() => void) | null;
+}
