+++ client/src/types/redux/IUserState.ts
@@ -1,0 +1,6 @@
+export interface IUserState {
+  userId: string;
+  userName: string;
+  userProfileImgPath: string;
+  isLogin: boolean;
+}
